chore(mockito): install android mockito
So TDD can be done on our own code.
However, need dependency injection or inversion of control for mock to be
injected in inner components. Thus never really usefull in our case because
doesn't achieve the objective which was to do TDD through system method call
analysis for notification. It's perhaps a bad idea overall to try to achieve
100% coverage. We should perhaps limit testing to logic and avoid TDD for the
rest.