fix(pheno): remove LOG CALL/SMS permissions
So app can be approved on google play due to new policies.

Due to a bug in google play console website, it wasn't possible
to break through the process even after having the perms removed.

The below potential fixes are the best approximation of a solution
but didn't work because
1. There is no box 'No, this release does not meet the SMS and
   Call log'.
2. There is 2 tracks with non compliant APK (see link 1's Edit
   sections).

Cf.
https://www.reddit.com/r/androiddev/comments/ajddj6/play_store_console_you_cant_edit_this_app_until/eeye216/
https://www.reddit.com/r/androiddev/comments/ajddj6/play_store_console_you_cant_edit_this_app_until/efkqmx2/

Still to try:
- Use Safari browser (not Chrome). cf. https://www.reddit.com/r/androiddev/comments/ajddj6/play_store_console_you_cant_edit_this_app_until/eff3imc/ (the guy looks pretty serious)
- Rename the package and republish as a different app.